# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira DC/Server Migration Audit API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JiraConnectionRequest {
  /** Base URL of the Jira DC/Server instance (e.g. https://jira.mycompany.com) */
  jiraUrl: string;
  /** Jira admin username */
  username: string;
  /** Jira admin password or personal access token */
  password: string;
  /** Optional friendly label for this audit */
  label?: string;
}

export type JiraConnectionResultDeploymentType =
  (typeof JiraConnectionResultDeploymentType)[keyof typeof JiraConnectionResultDeploymentType];

export const JiraConnectionResultDeploymentType = {
  server: "server",
  datacenter: "datacenter",
  cloud: "cloud",
  unknown: "unknown",
} as const;

export interface JiraConnectionResult {
  success: boolean;
  jiraVersion?: string;
  serverTitle?: string;
  deploymentType?: JiraConnectionResultDeploymentType;
  message: string;
}

export type AuditSessionStatus =
  (typeof AuditSessionStatus)[keyof typeof AuditSessionStatus];

export const AuditSessionStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface AuditSession {
  id: string;
  label?: string;
  jiraUrl: string;
  jiraVersion?: string;
  serverTitle?: string;
  deploymentType?: string;
  status: AuditSessionStatus;
  /** Migration readiness score 0-100 */
  readinessScore?: number;
  blockerCount?: number;
  warningCount?: number;
  infoCount?: number;
  createdAt: string;
  completedAt?: string | null;
  errorMessage?: string | null;
}

export type AuditFindingCategory =
  (typeof AuditFindingCategory)[keyof typeof AuditFindingCategory];

export const AuditFindingCategory = {
  system_info: "system_info",
  plugins: "plugins",
  custom_fields: "custom_fields",
  workflows: "workflows",
  permissions: "permissions",
  users: "users",
  data_volume: "data_volume",
  automation: "automation",
  projects: "projects",
} as const;

export type AuditFindingSeverity =
  (typeof AuditFindingSeverity)[keyof typeof AuditFindingSeverity];

export const AuditFindingSeverity = {
  blocker: "blocker",
  warning: "warning",
  info: "info",
} as const;

export interface AuditFinding {
  id: string;
  sessionId: string;
  category: AuditFindingCategory;
  severity: AuditFindingSeverity;
  title: string;
  description: string;
  count?: number | null;
  /** JSON-encoded extra details */
  details?: string | null;
  /** Recommended action for migration */
  migrationNote?: string | null;
}

export type AuditSessionDetail = AuditSession & {
  findings?: AuditFinding[];
};

export type AuditSummaryReadinessLevel =
  (typeof AuditSummaryReadinessLevel)[keyof typeof AuditSummaryReadinessLevel];

export const AuditSummaryReadinessLevel = {
  critical: "critical",
  at_risk: "at_risk",
  moderate: "moderate",
  good: "good",
  excellent: "excellent",
} as const;

export interface CategoryBreakdown {
  category: string;
  blockers: number;
  warnings: number;
  infos: number;
  /** Category-level score 0-100 */
  score: number;
}

export interface AuditSummary {
  sessionId: string;
  readinessScore: number;
  readinessLevel: AuditSummaryReadinessLevel;
  blockerCount: number;
  warningCount: number;
  infoCount: number;
  categoryBreakdown: CategoryBreakdown[];
  topBlockers: AuditFinding[];
}

export interface AuditStats {
  totalSessions: number;
  completedSessions: number;
  averageReadinessScore: number;
  totalBlockersFound: number;
  totalWarningsFound: number;
  mostCommonBlockerCategory?: string | null;
}