# lib/api-zod/src/generated/types/auditFinding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira DC/Server Migration Audit API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditFindingCategory } from "./auditFindingCategory";
import type { AuditFindingSeverity } from "./auditFindingSeverity";

export interface AuditFinding {
  id: string;
  sessionId: string;
  category: AuditFindingCategory;
  severity: AuditFindingSeverity;
  title: string;
  description: string;
  count?: number | null;
  /** JSON-encoded extra details */
  details?: string | null;
  /** Recommended action for migration */
  migrationNote?: string | null;
}