# lib/api-zod/src/generated/types/auditSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira DC/Server Migration Audit API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditSessionStatus } from "./auditSessionStatus";

export interface AuditSession {
  id: string;
  label?: string;
  jiraUrl: string;
  jiraVersion?: string;
  serverTitle?: string;
  deploymentType?: string;
  status: AuditSessionStatus;
  /** Migration readiness score 0-100 */
  readinessScore?: number;
  blockerCount?: number;
  warningCount?: number;
  infoCount?: number;
  createdAt: Date;
  completedAt?: Date | null;
  errorMessage?: string | null;
}