# lib/api-zod/src/generated/types/auditSessionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jira DC/Server Migration Audit API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditFinding } from "./auditFinding";
import type { AuditSession } from "./auditSession";

export type AuditSessionDetail = AuditSession & {
  findings?: AuditFinding[];
};